refactor: more and rename ShareLinks
