+++ src/components/TeacherComponents/LessonShareLinks/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./LessonShareLinks";
