+++ src/components/TeacherComponents/LessonShareLinks/linkConfig.ts
